Refactored to files in utils/messaging
Closes #17
